+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 # What's new:
-- Ported to 26.2
+- Improved accuracy
+- Added compatibility with Vivecraft
+- Improved the API so that registering modded projectiles is now actually possible
+- Rewrote a big part of the code